app/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
pydantic==2.6.3
pydantic-settings==2.2.1
chromadb>=0.4.0
redis>=5.0.0
httpx>=0.27.0
requests>=2.31.0
openai>=1.30.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
scikit-learn>=1.3.0
joblib>=1.3.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
python-docx>=1.1.0
